requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
transformers==4.36.2
torch==2.5.0
accelerate==0.25.0
pydantic==2.9.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==3.5.3
slowapi==0.1.5
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
PyYAML==6.0.1
requests==2.31.0
jsonschema==4.19.2